refactor(kademlia): remove unused code
